backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose
passlib[bcrypt]
bcrypt==4.0.1

python-dotenv